Made Profile Page v2 current version
