+++ src/views/ProfilePage.vue
@@ -1,16 +1,25 @@
-<script lang="ts" setup>
-import { useOnline, useScroll } from "@vueuse/core";
-import { ref, watch, computed } from "vue";
+<script setup lang="ts">
+import LocationPin from "../components/Icons/LocationPin.vue";
+import HeartIcon from "../components/Icons/HeartIcon.vue";
+import LinkIcon from "../components/Icons/LinkIcon.vue";
+import { useElementBounding } from "@vueuse/core";
+import { ref, watch, onMounted, computed } from "vue";
+import type { Blog, User } from "@/types";
 import router from "@/router";
-import type { RouteLocationRaw } from "vue-router";
-import { useQuery } from "@vue/apollo-composable";
-import type { Blog } from "@/types";
-import { GET_USER_BY_ID, GET_USER_BY_USERNAME } from "@/graphql";
 import { useAuth0 } from "@auth0/auth0-vue";
-import LoadingScreen from "@/components/LoadingScreen.vue";
-const isCompact = ref(false);
-const { y } = useScroll(window);
-const currentTab = ref<"posts" | "favorites">("posts");
+import {
+  GET_USER_BY_ID,
+  GET_USER_BY_USERNAME,
+  FOLLOW_USER,
+  UNFOLLOW_USER,
+} from "@/graphql";
+import { useMutation, useQuery } from "@vue/apollo-composable";
+import { useOnline } from "@vueuse/core";
+import PostItem from "../components/PostItem.vue";
+
+const online = useOnline();
+const nav = ref(null);
+
 const userParam = router.currentRoute.value.params.user;
 const { user: u } = useAuth0();
 const {
@@ -22,41 +31,29 @@ const {
 } = !isNaN(+userParam)
   ? useQuery(GET_USER_BY_ID, { id: userParam })
   : useQuery(GET_USER_BY_USERNAME, { username: userParam });
+
 const userFound = ref(false);
-const online = useOnline();
 onResult((r) => {
+  console.log(r.data);
   if (r.data.users.length == 0) {
     router.push("/404");
   } else {
     userFound.value = true;
   }
 });
-
 onError(() => {
-  console.error("Some Error Occured In Accessing DB! try to refetch");
+  console.error("Some Problem Occured! Please Refetch");
 });
+const { y } = useElementBounding(nav);
+const navIsCompact = ref(false);
+const tabs = ["About", "Posts", "Favourites", "Followers", "Followings"];
 
 //For user to change even when page not refreshed
 //For instance, checking a user's profile and then checking your own profile
 router.afterEach((to, from) => {
-  if (to.name === "users" && from.name === "users")
+  if (to.name === "users" && from.name === "users" && to.hash === "")
     refetch({ id: to.params.user, username: to.params.user });
 });
-watch(y, () => {
-  if (y.value > 20) {
-    isCompact.value = true;
-  } else {
-    isCompact.value = false;
-  }
-});
-
-function navigateTo(url: RouteLocationRaw) {
-  router.push(url);
-}
-
-function changeTab(tab: "posts" | "favorites") {
-  currentTab.value = tab;
-}
 
 const getFilteredBlogs = computed(() => {
   if (user.value.users[0].username === u.value?.nickname) {
@@ -64,195 +61,387 @@ const getFilteredBlogs = computed(() => {
   }
   return user.value.users[0].blogs.filter((blog: Blog) => blog.is_public);
 });
-</script>
+watch(y, () => {
+  //Will use different Method later
+  if (y.value === 80) {
+    navIsCompact.value = true;
+  } else {
+    navIsCompact.value = false;
+  }
+});
 
+function getFormattedDate(date: number) {
+  return Intl.DateTimeFormat(undefined, {
+    month: "long",
+    year: "numeric",
+  }).format(new Date(date));
+}
+
+function followUser(user: User) {
+  const { mutate } = useMutation(FOLLOW_USER);
+  mutate({ me: u.value.nickname, user: user.username });
+}
+function unfollowUser(user: User) {
+  const { mutate } = useMutation(UNFOLLOW_USER);
+  mutate({ me: u.value.nickname, user: user.username });
+}
+function isFollowed(user: any) {
+  for (let follower of user.follower_count.nodes || []) {
+    if (follower.user === u.value.nickname) {
+      return true;
+    }
+  }
+  return false;
+}
+const currentSection = ref<string | null>("About");
+onMounted(() => {
+  const observer = new IntersectionObserver(
+    ([entry]) => {
+      console.log("Observing " + entry);
+      if (entry.intersectionRatio > 0) {
+        currentSection.value = entry.target.getAttribute("id");
+      }
+    },
+    { threshold: 0, rootMargin: "0% 0px -75% 0px" }
+  );
+  document.querySelectorAll(".section .heading").forEach((section) => {
+    observer.observe(section);
+  });
+});
+</script>
 <template>
-  <main v-if="user && userFound && !loading" class="container">
-    <div class="user-display-container" :class="{ compact: isCompact }">
+  <main class="container" v-if="user && userFound && !loading">
+    <section class="user-info">
+      <div class="cover-image"></div>
       <img
         :src="user.users[0].profile_picture"
-        alt="user"
-        class="userPfp"
-        referrerpolicy="no-referrer"
+        alt="user-avatar"
+        class="avatar"
       />
-      <span class="fullname">{{ user.users[0].name }}</span>
-      <span class="username">@{{ user.users[0].username }}</span>
-    </div>
-    <span class="scrollDown" :class="{ compact: isCompact }">
-      <fa-icon icon="chevron-down" class="scrollDownIcon" />
-    </span>
-    <ul class="optionContainer">
-      <li
-        class="option"
-        :class="{ active: currentTab === 'posts' }"
-        @click="changeTab('posts')"
-      >
-        Posts
-      </li>
-      <li
-        class="option"
-        :class="{ active: currentTab === 'favorites' }"
-        @click="changeTab('favorites')"
-      >
-        Favorites
-      </li>
-    </ul>
-    <div class="main-data-container">
-      <div class="postsContainer" v-if="currentTab === 'posts'">
-        <div
-          class="post"
-          v-for="(blog, index) in getFilteredBlogs"
-          :key="blog.id"
-          @click="navigateTo(`/posts/${blog.id}`)"
+    </section>
+    <div class="background">
+      <section class="basic-user-info">
+        <div class="left-section">
+          <span class="full-name">{{ user.users[0].name }}</span>
+          <span
+            class="special-title"
+            v-for="(title, index) in user.users[0].special_title"
+            :key="index"
+            >{{ title }}</span
+          >
+          <span class="username">{{ user.users[0].username }}</span>
+        </div>
+        <div class="right-section">
+          <div class="analytics">
+            <div class="followers">
+              <span>Followers</span>
+              <span id="follower-count" class="count">
+                {{ user.users[0].follower_count.aggregate.count }}
+              </span>
+            </div>
+            <div class="followings">
+              <span>Followings</span>
+              <span id="followings-count" class="count">
+                {{ user.users[0].following_count.aggregate.count }}
+              </span>
+            </div>
+          </div>
+          <div
+            class="options"
+            v-if="u && u.nickname !== user.users[0].username"
+          >
+            <button
+              type="button"
+              class="follow-button"
+              @click="followUser(user.users[0])"
+              v-if="!isFollowed(user.users[0])"
+            >
+              Follow
+            </button>
+            <button
+              type="button"
+              class="unfollow-button"
+              @click="unfollowUser(user.users[0])"
+              v-else
+            >
+              Unfollow
+            </button>
+            <button type="button" class="message-button">Message</button>
+            <button type="button" class="report-button">Report</button>
+          </div>
+        </div>
+      </section>
+      <section class="user-detail-section card">
+        <span class="date-joined"
+          >Member since {{ getFormattedDate(user.users[0].created_at) }}</span
         >
-          <span class="blogItemIndex"> {{ index + 1 }}. </span>
-          <span class="blogTitle">
-            {{ blog.title }}
+        <div class="major-info">
+          <span class="location">
+            <LocationPin />
+            India
+          </span>
+          <span class="interests">
+            <HeartIcon />
+            Gaming, Coding
           </span>
         </div>
-        <div v-if="!getFilteredBlogs.length">Empty, just like my wallet</div>
-      </div>
-      <div
-        class="favoritePostsContainer"
-        v-else-if="currentTab === 'favorites'"
-      >
-        <div
-          class="post"
-          v-for="(blog, index) in user.users[0].liked_blogs"
-          :key="blog.id"
-          @click="navigateTo(`/posts/${blog.blog[0].id}`)"
-        >
-          <span class="blogItemIndex"> {{ index + 1 }}. </span>
-          <span class="blogTitle">
-            {{ blog.blog[0].title }}
-          </span>
+        <div class="social-media-links">
+          <a href="https://cryogon.netlify.app" class="website">
+            <LinkIcon />
+            cryogon.netlify.app
+          </a>
+          <!-- <a href="https://youtube.com/@cryogon" class="youtube">Youtube </a>
+          <a href="https://www.instagram.com/cryogonjs/" class="instagram"
+            >Instagram
+          </a> -->
         </div>
-        <div v-if="!user.users[0].liked_blogs.length">Nothing Here.</div>
-      </div>
+      </section>
+      <nav
+        class="tab-navigation"
+        :class="{ 'is-compact': navIsCompact }"
+        ref="nav"
+        id="nav"
+      >
+        <ul>
+          <router-link
+            v-for="(tab, i) in tabs"
+            :key="i"
+            :to="`#${tab}`"
+            class="tab-item"
+            :class="{ active: tab === currentSection }"
+            >{{ tab }}</router-link
+          >
+        </ul>
+      </nav>
+      <section
+        class="about-section card section"
+        ref="aboutSection"
+        id="aboutSection"
+      >
+        <h4 class="heading" id="About">About</h4>
+      </section>
+      <section
+        class="posts-section card section"
+        ref="postsSection"
+        id="postsSection"
+      >
+        <h4 class="heading" id="Posts">Posts</h4>
+        <PostItem
+          v-for="post in getFilteredBlogs"
+          :key="post.id"
+          :id="post.id"
+          :title="post.title"
+          :favorites="post.favourites.aggregate.count"
+          :posted_at="post.date_posted"
+          :visibility="post.is_public"
+        />
+      </section>
+      <section
+        class="favourite-section card section"
+        ref="favouritesSection"
+        id="favouritesSection"
+      >
+        <h4 class="heading" id="Favourites">Favourites</h4>
+      </section>
+      <section
+        class="follower-section card section"
+        ref="followersSection"
+        id="followersSection"
+      >
+        <h4 class="heading" id="Followers">Followers</h4>
+      </section>
+      <section
+        class="following-section card section"
+        ref="followingsSection"
+        id="followingsSection"
+      >
+        <h4 class="heading" id="Followings">Following</h4>
+      </section>
     </div>
   </main>
-  <main v-else-if="!online">You are currently offline</main>
-  <LoadingScreen v-else />
+  <main v-if="!online">Your are currently Offline</main>
 </template>
 <style scoped lang="scss">
 .container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding-block-start: 1rem;
-
-  .user-display-container {
-    display: grid;
+  transition: 100ms;
+  padding: 0 14vw;
+  .card {
+    min-height: 20rem;
+    margin-block-start: 1rem;
+    background-color: #303030;
+    border-radius: 12px;
+    padding: 0.3rem 1rem;
+    .heading {
+      font-weight: normal;
+      text-decoration: underline;
+    }
+  }
+  .user-info {
     position: relative;
-    align-self: flex-start;
-    margin-inline-start: 49%;
-    translate: -50% 0;
-    gap: 10px;
-    margin-bottom: 50rem;
-
-    transition: margin 0.5s;
-
-    .userPfp {
-      width: 20rem;
-      aspect-ratio: 1 / 1;
-      border-radius: 50%;
+    .cover-image {
+      height: 18rem;
+      width: 100%;
+      background-color: grey;
     }
-    .fullname {
-      font-size: 38px;
-    }
-    .fullname,
-    .username {
-      padding-inline-start: 2rem;
-    }
-    &.compact {
-      grid-template-columns: 9rem 1fr;
-      align-items: center;
-      margin-inline-start: 9vw;
-      margin-block: 5rem 3rem;
-      align-self: flex-start;
-      translate: 0;
-      animation: moveleft 0.4s ease-out;
-      @keyframes moveleft {
-      }
-      .userPfp {
-        width: 10rem;
-        grid-row: 1 / span 2;
-      }
-      .fullname {
-        font-size: 24px;
-        align-self: flex-end;
-      }
-      .username {
-        align-self: flex-start;
-      }
-    }
-  }
-  .scrollDown {
-    position: absolute;
-    bottom: 2rem;
-    animation: updown 500ms infinite alternate ease-out;
-    @keyframes updown {
-      from {
-        bottom: 1.8rem;
-      }
-      to {
-        bottom: 3rem;
-      }
-    }
-    &.compact {
-      display: none;
-    }
-    .scrollDownIcon {
-      scale: 2;
-    }
-  }
-  .optionContainer {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    width: 80%;
-    list-style: none;
-    gap: 5rem;
-    font-size: 20px;
-
-    .option {
-      cursor: pointer;
-    }
-    .active {
-      padding-inline: 1rem;
-      padding-block: 0.5rem;
+    .avatar {
+      width: 9rem;
+      height: 9rem;
+      display: block;
+      top: 12rem;
+      left: 2.5rem;
       border-radius: 2rem;
-      background-color: var(--link-hover-background);
+      position: absolute;
     }
   }
-  .main-data-container {
-    margin-block-start: 2rem;
-    width: 90%;
+  .background {
+    background-color: #202020;
     min-height: 30rem;
-    position: relative;
-    .postsContainer,
-    .favoritePostsContainer {
-      background-color: var(--card-background);
-      padding: 0.5rem;
-      .post {
-        padding: 0.4rem;
-        min-height: 3rem;
-        margin-block-end: 1rem;
-        &:hover {
-          background-color: var(--color-background);
-        }
-        .blogTitle {
+    padding: 1rem;
+    .basic-user-info {
+      display: flex;
+      flex-wrap: wrap;
+      padding-block-end: 1.5rem;
+      padding-inline: 2rem;
+      justify-content: space-between;
+      gap: 2rem;
+      .left-section {
+        display: flex;
+        flex-direction: column;
+        padding-block-start: 2.6rem;
+        margin-inline-end: 5rem;
+
+        .full-name {
           font-size: 20px;
-          letter-spacing: 0.3px;
-          cursor: pointer;
         }
-        .blogItemIndex {
-          font-size: 22px;
-          padding-inline: 1rem;
+        .special-title {
+          color: transparent;
+          background: linear-gradient(var(--tag-background));
+          background-clip: text;
+          -moz-background-clip: text;
+          -webkit-background-clip: text;
+        }
+        .username {
+          opacity: 0.6;
+        }
+      }
+      .right-section {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        .analytics {
+          display: flex;
+          gap: 4rem;
+          .followers,
+          .followings {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            .count {
+              color: transparent;
+              background: linear-gradient(var(--tag-background));
+              background-clip: text;
+              -moz-background-clip: text;
+              -webkit-background-clip: text;
+            }
+          }
+        }
+        .options {
+          display: flex;
+          gap: 1.5rem;
+          button {
+            padding: 0.3rem 0.9rem;
+            border-radius: 9px;
+            color: white;
+            outline-color: transparent;
+            border: 0;
+            border-bottom: 2px solid white;
+            margin-block-start: 2rem;
+            transition: box-shadow 300ms;
+            &:hover {
+              background-color: rgb(95, 145, 91);
+            }
+            &:active {
+              border-bottom: 0;
+            }
+          }
+          .follow-button,
+          .message-button {
+            background-color: var(--accent-color);
+          }
+          .unfollow-button {
+            background-color: grey;
+          }
+          .report-button {
+            background-color: #e75151;
+          }
         }
       }
     }
+    .user-detail-section {
+      min-height: auto;
+      .date-joined {
+        opacity: 0.6;
+        font-size: 14px;
+      }
+      .major-info {
+        display: flex;
+        gap: 10px;
+        margin-top: 0.1rem;
+        .location,
+        .interests {
+          display: flex;
+          align-items: center;
+          gap: 5px;
+        }
+      }
+      .social-media-links {
+        display: flex;
+        margin-block-start: 0.1rem;
+        flex-direction: column;
+        a {
+          display: flex;
+          align-items: center;
+          gap: 5px;
+          padding: 0;
+          &:hover {
+            text-decoration: underline;
+          }
+        }
+      }
+    }
+    .tab-navigation {
+      padding: 0.5rem;
+      padding-inline-start: 1rem;
+      margin-block-start: 1rem;
+      border-radius: 12px;
+      position: sticky;
+      z-index: 1;
+      top: 5rem;
+      background-color: #252525;
+      transition: 200ms;
+      &.is-compact {
+        background-color: #202020;
+        border-radius: 0;
+      }
+      .tab-item {
+        cursor: pointer;
+        position: relative;
+        &.active {
+          text-decoration: underline dotted;
+        }
+      }
+      ul {
+        display: flex;
+        list-style: none;
+        gap: 1rem;
+      }
+    }
+  }
+  @media screen and (max-width: 900px) {
+    padding: 0 4rem;
+  }
+  @media screen and (max-width: 700px) {
+    padding: 0 0;
   }
 }
 </style>
